+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 **Unreleased**
 --------------
 
+- **Fix:** Record lookups of generated static member inject functions for IC.
+- **Fix:** Dedupe merged overrides of `@Includes` accessors.
+
 0.3.6
 -----